requirements.txt
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
requests>=2.31.0
